+++ frontend/src/app/App.jsx
@@ -1,9 +1,10 @@
 // frontend/src/app/App.jsx
+import Header from "../components/Header";
 
 export default function App() {
     return (
         <>
-            <h1>Hello World</h1>
+            <Header />
         </>
     );
 }